+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Filter validation: Fix issue in Required filter validator with dot notation (#4221)
 * OpenAPI: Fix notice/warning for `response` without `content` in the `openapi_context` (#4210)
 * Serializer: Convert internal error to HTTP 400 in Ramsey uuid denormalization from invalid body string (#4200)
+* Symfony: Add tests with Symfony Uuid (#4230)
 
 ## 2.6.4